build: remove cjs build
Supporting cjs and esm builds are not neccessary. Esm builds should be enough for modern react
projects.

BREAKING CHANGE: Restructure the dist folder. The cjs build is removed and the esm build is moved to
the root of the dist folder. You need to change e.g. "@openpatch/patches/dist/esm/..." to
"@openpatch/patches/dist/...".