fix: add mjs extensions to files
For example Next.js was not able to determine the correct file by just looking at the package.json.
Therefore we needed to add mjs to als esm modules, which is also recommended by Node. In this
process we switched our build tool from tsc to esbuild to add the correct file-extension and to
speed the process up.Because we now fully support esm the @emotion/styled package was not working
anymore, thus we have reworked this too. Now everything should be working in CommonJS and
ES-Modules.