feat(deps): update major dependencies
We have updated patches to be compatible to React 18. We also updated all codemirror plugins, which
will not affect the use of patches. We also updated react-tabs this removes the export of
resetIdCounter, which is not necessary anymore. We also updated other dependencies aswell, which
also do not affect the use of patches.

BREAKING CHANGE: Export tabResetIdCounter is removed